+++ package.json
@@ -1,36 +1,46 @@
 {
-  "dependencies": {
-    "axios": "^1.6.5",
-    "express": "^4.18.2"
-  },
   "name": "quantum_krypto",
-  "description": "Crypto App",
-  "version": "1.0.0",
-  "main": "index.js",
-  "devDependencies": {
-    "coffee-script": "^1.12.7"
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@fontsource/roboto": "^5.0.8",
+    "@mui/icons-material": "^5.15.7",
+    "@mui/material": "^5.15.7",
+    "@mui/styled-engine-sc": "^6.0.0-alpha.14",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.3",
+    "react-scripts": "5.0.1",
+    "styled-components": "^6.1.8",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/SCORLEOs773/quantum_krypto.git"
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
-  "keywords": [
-    "Crypto",
-    "NFT",
-    "Portfolio",
-    "Forum",
-    "Quantum",
-    "Krypto",
-    "Hub",
-    "Cryptocurrency"
-  ],
-  "author": "Sukhad Sharma",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/SCORLEOs773/quantum_krypto/issues"
-  },
-  "homepage": "https://github.com/SCORLEOs773/quantum_krypto#readme"
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
