+++ src/components/QuantumHub.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import NFTHero from "./NFTHero";
+import PortfolioHero from "./PortfolioHero";
+
+export default function QuantumHub() {
+  return (
+    <div>
+      <NFTHero />
+      <PortfolioHero />
+    </div>
+  );
+}
