--- src/pages/BitBuzzNewsGroup.jsx
@@ -1,123 +1,0 @@
-import React from "react";
-
-const BitBuzzNewsGroup = () => {
-  return (
-    <div className="bit-buzz-news-group">
-      <div className="bit-buzz-news-group-child" />
-      <div className="bit-buzz-slide-frame">
-        <div className="slide-group">
-          <div className="news-item">
-            <div className="news-item-child" />
-            <b className="news-title">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon" alt="" src="/image-8@2x.png" />
-          </div>
-          <div className="news-item1">
-            <div className="news-item-item" />
-            <b className="news-title1">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container2">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon1" alt="" src="/image-8@2x.png" />
-          </div>
-          <div className="news-item2">
-            <div className="news-item-item" />
-            <b className="news-title2">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container4">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon2" alt="" src="/image-8@2x.png" />
-          </div>
-          <div className="news-item3">
-            <div className="news-item-child" />
-            <b className="news-title">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon" alt="" src="/image-8@2x.png" />
-          </div>
-          <div className="news-item4">
-            <div className="news-item-child" />
-            <b className="news-title">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon" alt="" src="/image-8@2x.png" />
-          </div>
-          <div className="news-item5">
-            <div className="news-item-item" />
-            <b className="news-title1">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container2">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon1" alt="" src="/image-8@2x.png" />
-          </div>
-          <div className="news-item6">
-            <div className="news-item-item" />
-            <b className="news-title1">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container2">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon1" alt="" src="/image-8@2x.png" />
-          </div>
-          <div className="news-item7">
-            <div className="news-item-item" />
-            <b className="news-title2">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container4">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon2" alt="" src="/image-8@2x.png" />
-          </div>
-          <div className="news-item8">
-            <div className="news-item-item" />
-            <b className="news-title2">NEWS TITLE</b>
-            <b className="lorem-ipsum-dolor-container4">
-              <span className="lorem-ipsum-dolor-container1">
-                <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-                <p className="join-the-quantum-community">&nbsp;</p>
-              </span>
-            </b>
-            <img className="image-8-icon2" alt="" src="/image-8@2x.png" />
-          </div>
-        </div>
-      </div>
-      <b className="stay-ahead-with-container">
-        <span className="lorem-ipsum-dolor-container1">
-          <span>{`Stay ahead with `}</span>
-          <span className="real-time">Real-Time</span>
-          <span>{` `}</span>
-          <span className="crypto-insights">CRYPTO INSIGHTS</span>
-        </span>
-      </b>
-      <b className="bit-buzz">BIT BUZZ</b>
-    </div>
-  );
-};
-
-export default BitBuzzNewsGroup;
