--- src/pages/Footer.jsx
@@ -1,45 +1,0 @@
-import React from "react";
-
-const Footer = () => {
-  return (
-    <div className="footer-home">
-      <div className="footer">
-        <div className="footer-child" />
-        <div className="footer-item" />
-        <b className="quantumkrypto">QuantumKrypto</b>
-        <b className="krypto">Krypto</b>
-        <b className="market">Market</b>
-        <b className="contact">Contact</b>
-        <b className="join-our-newsletter">Join Our Newsletter</b>
-        <div className="under-krypto-footer">
-          <div className="home">Home</div>
-          <div className="about">About</div>
-          <div className="buy-nfts">Buy NFTs</div>
-          <div className="sell-nfts">Sell NFTs</div>
-        </div>
-        <div className="under-market-footer">
-          <div className="browse-nfts">Browse NFTs</div>
-          <div className="buy-nfts1">Buy NFTs</div>
-          <div className="sell-nfts1">Sell NFTs</div>
-        </div>
-        <div className="under-contact-footer">
-          <div className="email">Email</div>
-          <div className="linkedin">LinkedIn</div>
-          <div className="buy-nfts">Instagram</div>
-          <div className="twitter">Twitter</div>
-        </div>
-        <div className="newsletter-email-box">
-          <div className="newsletter-email-box-child" />
-          <div className="email1">Email</div>
-          <div className="submit">
-            <div className="submit-child" />
-            <b className="submit1">SUBMIT</b>
-          </div>
-        </div>
-        <img className="logo-footer-icon" alt="" src="/logo-footer@2x.png" />
-      </div>
-    </div>
-  );
-};
-
-export default Footer;
