--- src/pages/PortfolioHero.jsx
@@ -1,23 +1,0 @@
-import React from "react";
-
-const PortfolioHero = () => {
-  return (
-    <div className="portfolio-hero">
-      <div className="portfolio-hero-child" />
-      <b className="built-in-nft-analytics">
-        Built-In NFT Analytics for Portfolio Mastery
-      </b>
-      <b className="harness-insights-with">
-        Harness insights with our integrated analytics dashboard; track and
-        optimize the value of your Quantum Krypto portfolio effortlessly.
-      </b>
-      <div className="view-pricing1">
-        <b className="get-quantum-prime-container">View PLANS</b>
-      </div>
-      <img className="sapiens-icon" alt="" src="/sapiens.svg" />
-      <b className="portfolio">PORTFOLIO</b>
-    </div>
-  );
-};
-
-export default PortfolioHero;
