--- src/pages/Testimonial.jsx
@@ -1,51 +1,0 @@
-import React from "react";
-
-const TestimonialSection = () => {
-  return (
-    <div className="testimonial-section">
-      <b className="testimonials">TESTIMONIALS</b>
-      <b className="read-what-others-container">
-        <p className="join-the-quantum-community">{`Read What Others `}</p>
-        <p className="join-the-quantum-community">Have To Say</p>
-      </b>
-      <div className="testimonial">
-        <div className="testimonial-child" />
-        <b className="evan-white">EVAN WHITE</b>
-        <img className="mask-group-icon" alt="" src="/mask-group@2x.png" />
-        <b className="lorem-ipsum-dolor-container18">
-          <span className="lorem-ipsum-dolor-container1">
-            <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-            <p className="join-the-quantum-community">&nbsp;</p>
-          </span>
-        </b>
-        <div className="testimonial-item" />
-      </div>
-      <div className="testimonial1">
-        <div className="testimonial-child" />
-        <b className="evan-white">EVAN WHITE</b>
-        <img className="mask-group-icon" alt="" src="/mask-group@2x.png" />
-        <b className="lorem-ipsum-dolor-container18">
-          <span className="lorem-ipsum-dolor-container1">
-            <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-            <p className="join-the-quantum-community">&nbsp;</p>
-          </span>
-        </b>
-        <div className="testimonial-item" />
-      </div>
-      <div className="testimonial2">
-        <div className="testimonial-child" />
-        <b className="evan-white">EVAN WHITE</b>
-        <img className="mask-group-icon" alt="" src="/mask-group@2x.png" />
-        <b className="lorem-ipsum-dolor-container18">
-          <span className="lorem-ipsum-dolor-container1">
-            <p className="join-the-quantum-community">{`Lorem ipsum dolor sit amet. Sit rerum soluta est quisquam enim in quia illum sed consequatur magni non dolor saepe. `}</p>
-            <p className="join-the-quantum-community">&nbsp;</p>
-          </span>
-        </b>
-        <div className="testimonial-item" />
-      </div>
-    </div>
-  );
-};
-
-export default TestimonialSection;
